+++ src/postorius/tests/mailman_api_tests/test_list_summary.py
@@ -20,6 +20,7 @@ from django.contrib.auth.models import User
 from django.urls import reverse
 
 from allauth.account.models import EmailAddress
+from django_mailman3.lib.mailman import get_mailman_user
 
 from postorius.forms import ListAnonymousSubscribe
 from postorius.tests.utils import ViewTestCase
@@ -194,3 +195,35 @@ def function:
         response = self.client.get(reverse('list_summary',
                                            args=('foo@example.com',)))
         self.assertContains(response, '<pre><code>def function:')
+
+    def test_list_summary_already_subscribed(self):
+        self.foo_list.subscribe(
+            'test@example.com',
+            pre_verified=True, pre_confirmed=True, pre_approved=True)
+        self.client.login(username='testuser', password='testpass')
+        response = self.client.get(reverse('list_summary',
+                                           args=('foo@example.com', )))
+        # Assert two parts separately due to sometimes being newlines and stuff
+        # in the response HTML.
+        self.assertContains(response,
+                            'You are subscribed to this list with the '
+                            'following address:')
+        self.assertContains(response, ' <em>test@example.com</em>')
+
+    def test_list_summary_already_subscribed_user(self):
+        mm_user = get_mailman_user(self.user)
+        mm_user.addresses[0].verify()
+        mm_user.preferred_address = self.user.email
+        self.foo_list.subscribe(
+            str(mm_user.user_id),
+            pre_verified=True, pre_confirmed=True, pre_approved=True)
+        self.client.login(username='testuser', password='testpass')
+        response = self.client.get(reverse('list_summary',
+                                           args=('foo@example.com', )))
+        # Assert two parts separately due to sometimes being newlines and stuff
+        # in the response HTML.
+        self.assertContains(response,
+                            'You are subscribed to this list with the '
+                            'following address:')
+        self.assertContains(response,
+                            ' <em>Primary Address (test@example.com)</em>')